Add GameManagerSystem
It counts the number of entities that have TankPlayer components (i.e. tanks).
It's assumed that when a tank is destroyed, its entity gets the Disabled component
added. (All tanks are de-Disabled at the beginning of each round.)

If only one tank is left, the TankPlayer.Score is incremented.
If the round winner now has enough points to win the game, then
the game is over. Otherwise, begin a new round (currently a no-op).

If zero are left, the round is a draw and a new round is started.

Still TODO: spawn/activate tanks, perform other new-round/new-game logic.