Better workaround for particle system stall.
The actual problem was that the particle system update was blocking until
the FollowSystem jobs finished, which in turn was blocking until the
transform system jobs finished. Now we just run the FollowSystem before
the transform system, which means it's following the previous frame's
transform, but the bubble is gone.